fix(embedding): serialize FloatArray to BLOB before DAO write
Room's @Query treats FloatArray as a collection, generating 384
separate ? placeholders instead of a single BLOB binding. Pre-serialize
to ByteArray in the worker using the same little-endian layout as
AppTypeConverters so the column round-trips correctly on read.

+++ core-data/src/main/java/com/github/maskedkunisquat/wulfpak/core/data/dao/ActivityDao.kt
@@ -65,5 +65,5 @@ interface ActivityDao {
     suspend fun getParticipantsByPerson(personId: UUID): List<ActivityParticipant>
 
     @Query("UPDATE activities SET embedding = :embedding WHERE id = :id")
-    suspend fun updateEmbedding(id: UUID, embedding: FloatArray)
+    suspend fun updateEmbedding(id: UUID, embedding: ByteArray)
 }
